+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/src/models/mappers.ts
@@ -1737,6 +1737,20 @@ export const MigrationItemProperties: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      recoveryServicesProviderId: {
+        serializedName: "recoveryServicesProviderId",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      replicationStatus: {
+        serializedName: "replicationStatus",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
       migrationState: {
         serializedName: "migrationState",
         readOnly: true,
@@ -1760,6 +1774,20 @@ export const MigrationItemProperties: coreClient.CompositeMapper = {
       },
       lastTestMigrationStatus: {
         serializedName: "lastTestMigrationStatus",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      lastMigrationTime: {
+        serializedName: "lastMigrationTime",
+        readOnly: true,
+        type: {
+          name: "DateTime"
+        }
+      },
+      lastMigrationStatus: {
+        serializedName: "lastMigrationStatus",
         readOnly: true,
         type: {
           name: "String"
@@ -2095,6 +2123,91 @@ export const MigrationRecoveryPointProperties: coreClient.CompositeMapper = {
       recoveryPointType: {
         serializedName: "recoveryPointType",
         readOnly: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const PauseReplicationInput: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "PauseReplicationInput",
+    modelProperties: {
+      properties: {
+        serializedName: "properties",
+        type: {
+          name: "Composite",
+          className: "PauseReplicationInputProperties"
+        }
+      }
+    }
+  }
+};
+
+export const PauseReplicationInputProperties: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "PauseReplicationInputProperties",
+    modelProperties: {
+      instanceType: {
+        serializedName: "instanceType",
+        required: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const ResumeReplicationInput: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ResumeReplicationInput",
+    modelProperties: {
+      properties: {
+        serializedName: "properties",
+        type: {
+          name: "Composite",
+          className: "ResumeReplicationInputProperties"
+        }
+      }
+    }
+  }
+};
+
+export const ResumeReplicationInputProperties: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ResumeReplicationInputProperties",
+    modelProperties: {
+      providerSpecificDetails: {
+        serializedName: "providerSpecificDetails",
+        type: {
+          name: "Composite",
+          className: "ResumeReplicationProviderSpecificInput"
+        }
+      }
+    }
+  }
+};
+
+export const ResumeReplicationProviderSpecificInput: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ResumeReplicationProviderSpecificInput",
+    uberParent: "ResumeReplicationProviderSpecificInput",
+    polymorphicDiscriminator: {
+      serializedName: "instanceType",
+      clientName: "instanceType"
+    },
+    modelProperties: {
+      instanceType: {
+        serializedName: "instanceType",
+        required: true,
         type: {
           name: "String"
         }
@@ -11926,8 +12039,22 @@ export const VMwareCbtProtectedDiskDetails: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      seedBlobUri: {
+        serializedName: "seedBlobUri",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
       targetManagedDiskId: {
         serializedName: "targetManagedDiskId",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      targetBlobUri: {
+        serializedName: "targetBlobUri",
         readOnly: true,
         type: {
           name: "String"
@@ -12000,6 +12127,30 @@ export const VMwareCbtNicDetails: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      testNetworkId: {
+        serializedName: "testNetworkId",
+        type: {
+          name: "String"
+        }
+      },
+      testSubnetName: {
+        serializedName: "testSubnetName",
+        type: {
+          name: "String"
+        }
+      },
+      testIPAddress: {
+        serializedName: "testIPAddress",
+        type: {
+          name: "String"
+        }
+      },
+      testIPAddressType: {
+        serializedName: "testIPAddressType",
+        type: {
+          name: "String"
+        }
+      },
       targetNicName: {
         serializedName: "targetNicName",
         type: {
@@ -12058,6 +12209,18 @@ export const VMwareCbtNicInput: coreClient.CompositeMapper = {
         type: {
           name: "String"
         }
+      },
+      testSubnetName: {
+        serializedName: "testSubnetName",
+        type: {
+          name: "String"
+        }
+      },
+      testStaticIPAddress: {
+        serializedName: "testStaticIPAddress",
+        type: {
+          name: "String"
+        }
       }
     }
   }
@@ -12077,6 +12240,12 @@ export const VMwareCbtUpdateDiskInput: coreClient.CompositeMapper = {
       },
       targetDiskName: {
         serializedName: "targetDiskName",
+        type: {
+          name: "String"
+        }
+      },
+      isOSDisk: {
+        serializedName: "isOSDisk",
         type: {
           name: "String"
         }
@@ -13861,6 +14030,13 @@ export const VMwareCbtMigrationDetails: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      storageAccountId: {
+        serializedName: "storageAccountId",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
       targetVmName: {
         serializedName: "targetVmName",
         type: {
@@ -13935,6 +14111,12 @@ export const VMwareCbtMigrationDetails: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      testNetworkId: {
+        serializedName: "testNetworkId",
+        type: {
+          name: "String"
+        }
+      },
       vmNics: {
         serializedName: "vmNics",
         type: {
@@ -13996,6 +14178,13 @@ export const VMwareCbtMigrationDetails: coreClient.CompositeMapper = {
           name: "Number"
         }
       },
+      resumeProgressPercentage: {
+        serializedName: "resumeProgressPercentage",
+        readOnly: true,
+        type: {
+          name: "Number"
+        }
+      },
       initialSeedingRetryCount: {
         serializedName: "initialSeedingRetryCount",
         readOnly: true,
@@ -14005,6 +14194,13 @@ export const VMwareCbtMigrationDetails: coreClient.CompositeMapper = {
       },
       resyncRetryCount: {
         serializedName: "resyncRetryCount",
+        readOnly: true,
+        type: {
+          name: "Number"
+        }
+      },
+      resumeRetryCount: {
+        serializedName: "resumeRetryCount",
         readOnly: true,
         type: {
           name: "Number"
@@ -14130,8 +14326,20 @@ export const VMwareCbtEnableMigrationInput: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      testNetworkId: {
+        serializedName: "testNetworkId",
+        type: {
+          name: "String"
+        }
+      },
       targetSubnetName: {
         serializedName: "targetSubnetName",
+        type: {
+          name: "String"
+        }
+      },
+      testSubnetName: {
+        serializedName: "testSubnetName",
         type: {
           name: "String"
         }
@@ -14256,6 +14464,12 @@ export const VMwareCbtUpdateMigrationItemInput: coreClient.CompositeMapper = {
           name: "String"
         }
       },
+      testNetworkId: {
+        serializedName: "testNetworkId",
+        type: {
+          name: "String"
+        }
+      },
       vmNics: {
         serializedName: "vmNics",
         type: {
@@ -14344,6 +14558,26 @@ export const VMwareCbtMigrateInput: coreClient.CompositeMapper = {
   }
 };
 
+export const VMwareCbtResumeReplicationInput: coreClient.CompositeMapper = {
+  serializedName: "VMwareCbt",
+  type: {
+    name: "Composite",
+    className: "VMwareCbtResumeReplicationInput",
+    uberParent: "ResumeReplicationProviderSpecificInput",
+    polymorphicDiscriminator:
+      ResumeReplicationProviderSpecificInput.type.polymorphicDiscriminator,
+    modelProperties: {
+      ...ResumeReplicationProviderSpecificInput.type.modelProperties,
+      deleteMigrationResources: {
+        serializedName: "deleteMigrationResources",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const VMwareCbtResyncInput: coreClient.CompositeMapper = {
   serializedName: "VMwareCbt",
   type: {
@@ -14387,6 +14621,18 @@ export const VMwareCbtTestMigrateInput: coreClient.CompositeMapper = {
         required: true,
         type: {
           name: "String"
+        }
+      },
+      vmNics: {
+        serializedName: "vmNics",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "VMwareCbtNicInput"
+            }
+          }
         }
       }
     }
@@ -19014,14 +19260,12 @@ export const VMwareCbtContainerMappingInput: coreClient.CompositeMapper = {
       ...ReplicationProviderSpecificContainerMappingInput.type.modelProperties,
       keyVaultId: {
         serializedName: "keyVaultId",
-        required: true,
         type: {
           name: "String"
         }
       },
       keyVaultUri: {
         serializedName: "keyVaultUri",
-        required: true,
         type: {
           name: "String"
         }
@@ -19035,14 +19279,12 @@ export const VMwareCbtContainerMappingInput: coreClient.CompositeMapper = {
       },
       storageAccountSasSecretName: {
         serializedName: "storageAccountSasSecretName",
-        required: true,
         type: {
           name: "String"
         }
       },
       serviceBusConnectionStringSecretName: {
         serializedName: "serviceBusConnectionStringSecretName",
-        required: true,
         type: {
           name: "String"
         }
@@ -21526,6 +21768,7 @@ export let discriminators = {
   EnableMigrationProviderSpecificInput: EnableMigrationProviderSpecificInput,
   UpdateMigrationItemProviderSpecificInput: UpdateMigrationItemProviderSpecificInput,
   MigrateProviderSpecificInput: MigrateProviderSpecificInput,
+  ResumeReplicationProviderSpecificInput: ResumeReplicationProviderSpecificInput,
   ResyncProviderSpecificInput: ResyncProviderSpecificInput,
   TestMigrateProviderSpecificInput: TestMigrateProviderSpecificInput,
   ConfigurationSettings: ConfigurationSettings,
@@ -21600,6 +21843,7 @@ export let discriminators = {
   "EnableMigrationProviderSpecificInput.VMwareCbt": VMwareCbtEnableMigrationInput,
   "UpdateMigrationItemProviderSpecificInput.VMwareCbt": VMwareCbtUpdateMigrationItemInput,
   "MigrateProviderSpecificInput.VMwareCbt": VMwareCbtMigrateInput,
+  "ResumeReplicationProviderSpecificInput.VMwareCbt": VMwareCbtResumeReplicationInput,
   "ResyncProviderSpecificInput.VMwareCbt": VMwareCbtResyncInput,
   "TestMigrateProviderSpecificInput.VMwareCbt": VMwareCbtTestMigrateInput,
   "ConfigurationSettings.HyperVVirtualMachine": HyperVVirtualMachineDetails,